requirements.txt
pandas
Gradio
datasets
evaluate
scikit-learn
transformers
torch
torchvision
transformers[torch]
python-dotenv
langchain
langchain_community
sentence-transformers
pinecone
# langchain-community-unstructured
scrapy
beautifulsoup4
pinecone[grpc]
lightning
fastapi